refactored namespace to org.wordpress.mobile
