+++ android/src/main/AndroidManifest.xml
@@ -16,5 +16,5 @@
 -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.android.recyclerview">
+    package="org.wordpress.mobile.ReactNativeAztec">
 </manifest>
