+++ android/src/main/res/layout/aztec_main.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<com.example.android.ReactAztec.ReactAztecView xmlns:android="http://schemas.android.com/apk/res/android"
+<org.wordpress.mobile.ReactNativeAztec.ReactAztecView xmlns:android="http://schemas.android.com/apk/res/android"
                 xmlns:aztec="http://schemas.android.com/apk/res-auto"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent">
@@ -17,7 +17,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" >
 
-        <com.example.android.ReactAztec.ReactAztecText
+        <org.wordpress.mobile.ReactNativeAztec.ReactAztecText
             android:id="@+id/aztec"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -53,4 +53,4 @@
             aztec:codeTextColor="@android:color/black"/>
 
     </FrameLayout>
-</com.example.android.ReactAztec.ReactAztecView>
+</org.wordpress.mobile.ReactNativeAztec.ReactAztecView>
