+++ example/android/app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
 -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.android.recyclerview"
+    package="com.example.android"
     android:versionCode="1"
     android:versionName="1.0">
