+++ example/android/app/src/main/java/com/example/android/common/activities/SampleRNBaseActivity.java
@@ -19,8 +19,8 @@ package com.example.android.common.activities;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
 
-import com.example.android.ReactAztec.ReactAztecPackage;
-import com.example.android.recyclerview.BuildConfig;
+import org.wordpress.mobile.ReactNativeAztec.ReactAztecPackage;
+import org.wordpress.mobile.ReactNativeAztec.BuildConfig;
 import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.common.LifecycleState;
 import com.facebook.react.modules.core.DefaultHardwareBackBtnHandler;
